u-boot(imx8q_evk): prepare build_info.h
When preparing u-boot within the NXP' mkimage tool a reference to the original
git repository is implicitly needed to copy some commit id into a build_info
header. As we download the u-boot port as tarball snapshot from github, the
corresponding git information is missing, which leads to failures of the mkimage
tool. Therefore, we create a dummy build_info header explicitely.