framebuffer/imx8mq: use linux-imx 6.1.20
This commit introduces an updated driver that uses the current
linux-imx (6.1.20) version.

It also contains a unified test run-script that works on both the
EVK as well as the MNT Reform2. The HDMI port on the later device
can be exercised by changing the device-tree route to
'imx8mq_fb-mnt_reform2-hdmi.dtb'.