platform/imx8mp: service watchdog timer
Once the watchdog timer is enabled (e.g., by the boot loader) it cannot
be stopped. In order to not trigger a sytem reset when the watchdog
fires, we have to write a service sequence upon a watchdog-timer
interrupt (default is 2s before timeout). We only enable this behavior
if the watchdog timer is already running.

issue genodelabs/genode#5264